# docs/swagger.yml
swagger: "2.0"
info:
  description: "Ensuring access to notarized metadata"
  version: "1.0.0"
  title: "Pando"
  contact:
    name: KenLabs
    url: https://join.slack.com/t/ken-labs/shared_invite/zt-ylfrwtl2-H4ZyLCzw94wfiuC7VtAy~A
  license:
    name: "Apache 2.0 and MIT"
    url: "https://github.com/pando-project/pando"
host: "pando.kencloud.com:9011"
basePath: "/"


tags:
- name: "pando"
  description: "Get Pando's details information"
- name: "provider"
  description: "Register provider and get info of a specific provider"
- name: "metadata"
  description: "Get metadata details information"


schemes:
- "https"
- "http"


paths:
  /pando/subscribe:
    get:
      tags:
      - "pando"
      summary: "Let Pando subscribe a topic for data sync with a specific provider"
      description: ""
      operationId: "subscribeProvider"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "provider"
        type: "string"
        description: "PeerID of a provider to subscribe"
        required: true
      responses:
        "201":
          description: "Register success"
          schema: 
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              code: 201
              message: "opration success"
              data: {}
        "400":
          description: "Invalid request"
          schema:
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              code: 400
              message: "invalid peerID"
              data: {}
  /pando/info:
    get:
      tags:
      - "pando"
      summary: "Get Pando's peerID and multiaddress"
      description: ""
      operationId: "getPandoInfo"
      produces:
      - "application/json"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "ok"
              code: 200
              data:
                peerID: "12D3KooWEpehzfFV4U1SQFcWEXBkP3zseGszEGAPfApS9as2Swgi"
                APIAddresses:
                  GRAPHQL_API: "/ip4/1.1.1.1/tcp/8002"
                  GRAPHSYNC_API: "/ip4/1.1.1.1/tcp/8003"
                  HTTP_API: "/ip4/1.1.1.1/tcp/8001"

  /pando/metrics:
    get:
      tags:
      - "pando"
      summary: "Prometheus exporter page of Pando"
      description: ""
      operationId: "getPandoMetrics"
      produces:
      - "application/json"
      responses:
        "200":
          description: "OK"
  /pando/health:
    options:
      tags:
      - "pando"
      summary: "Check if Pando is alive"
      description: ""
      operationId: "checkPandoAlive"
      responses:
        "200":
          description: "OK"
    
  /provider/register:
    post:
      tags:
      - "provider"
      summary: "Register a new provider in Pando"
      description: ""
      operationId: "registerProvider"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Enveloped provider registration protobuff info"
        required: true
        schema:
          $ref: "#/definitions/Provider"
      responses:
        "201":
          description: "Register success"
          schema: 
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "opration success"
              data: {}
        "400":
          description: "Invalid request"
          schema:
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "invalid private key"
              data: {}
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "internal server error: <error message>"
              data: {}
  /provider/info:
    get:
      tags:
        - provider
      summary: "Get Providers' PeerID, Multiaddress and Miner address"
      description: ""
      operationId: "getProviderInfo"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "peerid"
          type: "string"
          description: "get a specific provider's info with its peerid, if empty, return all providers' info"
          required: false
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "ok"
              code: 200
              data:
                "registeredProviders":
                  "12D3KooWMm4sgwMsbzdGnLNhQv4dgMvqyp2JAAPHJHtRWVvjG8rn":
                    "MultiAddr": []
                    "MinerAddr": []
                "unregisteredProviders": {}
  /provider/head:
    get:
      tags:
        - provider
      summary: "Get the head cid of a specific provider"
      description: ""
      operationId: "getProviderHead"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "peerid"
          type: "string"
          description: "Get the head cid of a specific provider"
          required: true
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "ok"
              code: 200
              data:
                Cid: "baguqeeqqisoxg5itsdg5inuixczplgymd4"

  /metadata/list:
    get:
      tags:
      - "metadata"
      summary: "get a list involved all the metadata snapshot cids"
      description: ""
      operationId: "listMetadataSnapshots"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "ok"
          schema: 
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "operation success"
              data:
                metadata:
                - "cid1"
                - "cid2"
  /metadata/snapshot:
    get:
      tags:
      - "metadata"
      summary: "get a snapshot detail, can queried by its height or cid"
      description: ""
      operationId: "getMetadataSnapshotInfo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "query"
          name: "height"
          type: string
          description: "Height of a snapshot to retrieve"
          required: false
        - in: query
          name: "cid"
          type: string
          description: "Cid of a snapshot to retrieve"
          required: false
      responses:
        "200":
          description: "ok"
          schema: 
            $ref: "#/definitions/APIResponse"
          examples:
            application/json:
              message: "operation success"
              data:
                metadata:
                - "cid1"
                - "cid2"
              
definitions:
  Provider:
    type: object
    properties:
      cid: 
        type: string
      envelop: 
        type: string
  APIResponse:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
      data:
        type: object
externalDocs:
  description: "Find out more about Pando"
  url: "https://github.com/pando-project/pando"